refactor: split windows specific code out of hotkeywrapper
